backend-temp/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
filelock==3.14.0
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.4
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.29.0
packaging==24.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.0
setuptools==70.0.0
sniffio==1.3.1
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
